add set bnchmarks
